Renamed LogWatcher to LogcatWatcher
